+++ CotEditor/Sources/AppDelegate.swift
@@ -90,7 +90,9 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
     override init() {
         
         // register default setting values
-        let defaults = DefaultSettings.defaults.mapKeys(\.rawValue)
+        let defaults = DefaultSettings.defaults
+            .compactMapValues { $0 }
+            .mapKeys(\.rawValue)
         UserDefaults.standard.register(defaults: defaults)
         NSUserDefaultsController.shared.initialValues = defaults
         
@@ -138,7 +140,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
         SyntaxManager.shared.$settingNames
             .map { $0.map { NSMenuItem(title: $0, action: #selector((any SyntaxChanging).changeSyntax), keyEquivalent: "") } }
             .receive(on: RunLoop.main)
-            .sink { [weak self] (items) in
+            .sink { [weak self] items in
                 guard let menu = self?.syntaxesMenu else { return }
                 
                 let recolorItem = menu.items.first { $0.action == #selector((any SyntaxChanging).recolorAll) }
@@ -165,7 +167,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
         // build Unicode normalization menu items
         self.normalizationMenu?.items = (UnicodeNormalizationForm.standardForms + [nil] +
                                          UnicodeNormalizationForm.modifiedForms)
-            .map { (form) in
+            .map { form in
                 guard let form else { return .separator() }
                 
                 let item = NSMenuItem()
@@ -329,6 +331,13 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
     @IBAction func showSettingsWindow(_ sender: Any?) {
         
         self.settingsWindowController.showWindow(sender)
+    }
+    
+    
+    /// Show the Quick Action command bar.
+    @IBAction func showQuickActions(_ sender: Any?) {
+        
+        CommandBarWindowController.shared.showWindow(sender)
     }