+++ CotEditor/Sources/EditorTextViewController.swift
@@ -28,7 +28,7 @@ import AppKit
 import Combine
 import SwiftUI
 
-final class EditorTextViewController: NSViewController, NSTextViewDelegate {
+final class EditorTextViewController: NSViewController, NSServicesMenuRequestor, NSTextViewDelegate {
     
     // MARK: Enums
     
@@ -93,7 +93,7 @@ final class EditorTextViewController: NSViewController, NSTextViewDelegate {
         
         // observe text orientation for line number view
         self.orientationObserver = self.textView!.publisher(for: \.layoutOrientation, options: .initial)
-            .sink { [weak self] (orientation) in
+            .sink { [weak self] orientation in
                 guard let self else { return assertionFailure() }
                 
                 self.stackView?.orientation = switch orientation {
@@ -109,7 +109,7 @@ final class EditorTextViewController: NSViewController, NSTextViewDelegate {
         self.writingDirectionObserver = self.textView!.publisher(for: \.baseWritingDirection)
             .removeDuplicates()
             .map { $0 == .rightToLeft }
-            .sink { [weak self] (isRTL) in
+            .sink { [weak self] isRTL in
                 guard
                     let stackView = self?.stackView,
                     let lineNumberView = self?.lineNumberView
@@ -154,6 +154,41 @@ final class EditorTextViewController: NSViewController, NSTextViewDelegate {
         if coder.decodeBool(forKey: SerializationKey.showsAdvancedCounter) {
             self.showAdvancedCharacterCounter()
         }
+    }
+    
+    
+    
+    // MARK: View Controller
+    
+    override func validRequestor(forSendType sendType: NSPasteboard.PasteboardType?, returnType: NSPasteboard.PasteboardType?) -> Any? {
+        
+        // accept continuity camera
+        //   - Take Photo: .jpeg, .tiff
+        //   - Scan Documents: .pdf, .tiff
+        //   - Sketch: .png
+        if let returnType, NSImage.imageTypes.contains(returnType.rawValue) {
+            return (returnType != .png) ? self : nil
+        }
+        
+        return super.validRequestor(forSendType: sendType, returnType: returnType)
+    }
+    
+    
+    
+    // MARK: Services Menu Requestor
+    
+    func readSelection(from pboard: NSPasteboard) -> Bool {
+        
+        // scan from continuity camera
+        if pboard.canReadItem(withDataConformingToTypes: NSImage.imageTypes),
+           let image = NSImage(pasteboard: pboard)
+        {
+            self.popoverLiveText(image: image)
+            
+            return true
+        }
+        
+        return false
     }
     
     
@@ -219,7 +254,7 @@ final class EditorTextViewController: NSViewController, NSTextViewDelegate {
         let lineCount = (string as NSString).substring(with: textView.selectedRange).numberOfLines
         let lineRange = FuzzyRange(location: lineNumber, length: lineCount)
         
-        let view = GoToLineView(lineRange: lineRange) { (lineRange) in
+        let view = GoToLineView(lineRange: lineRange) { lineRange in
             guard let range = textView.string.rangeForLine(in: lineRange) else { return false }
             
             textView.select(range: range)
@@ -238,7 +273,7 @@ final class EditorTextViewController: NSViewController, NSTextViewDelegate {
         
         guard let textView = self.textView else { return assertionFailure() }
         
-        let view = UnicodeInputView { [unowned textView] (character) in
+        let view = UnicodeInputView { [unowned textView] character in
             // flag to skip line ending sanitization
             textView.isApprovedTextChange = true
             defer { textView.isApprovedTextChange = false }
@@ -310,6 +345,28 @@ final class EditorTextViewController: NSViewController, NSTextViewDelegate {
     
     
     // MARK: Private Methods
+    
+    /// Show a popover indicating the given image and live text detection.
+    ///
+    /// - Parameter image: The image to scan text.
+    private func popoverLiveText(image: NSImage) {
+        
+        guard let textView = self.textView else { return assertionFailure() }
+        
+        let rootView = LiveTextInsertionView(image: image) { [weak textView] string in
+            guard let textView else { return }
+            textView.replace(with: string, range: textView.selectedRange, selectedRange: nil)
+        }
+        let viewController = NSHostingController(rootView: rootView)
+        viewController.sizingOptions = .preferredContentSize
+        viewController.rootView.parent = viewController
+        
+        let positioningRect = textView.boundingRect(for: textView.selectedRange)?.insetBy(dx: -1, dy: -1) ?? .zero
+        
+        textView.scrollRangeToVisible(textView.selectedRange)
+        self.present(viewController, asPopoverRelativeTo: positioningRect, of: textView, preferredEdge: .maxY, behavior: .transient)
+    }
+    
     
     /// Hide existing advanced character counter.
     ///